tests/pp/pp-def-7.c: add a case which fails due to incorrect expansion
We are still not correctly expanding macro arguments. Instead of
expanding them separately first, like we should, we just copy them and
let them be expanded later when encountered. That mostly works
except in some cases, notably recursive macro invocation.

NEG(NEG(1)) gets *incorrectly* expanded to (-(NEG(1))), which incorrectly
causes us to assume that this is a recursive expansion.

This commit adds a test case demonstrating the failure in anticipation of
fixing it.