CParser.m4.y: add support for C-style implicit function declarations
'primary-expression' already reports undefined identifiers, so by the
the time we get to 'postfix-expression' and figure out it is a function
call and an implicit function declaration, it is too late.

So, we need to directly handle 'identifier "(" ...' in 'postfix-
expression'. The straight forward implementation, which we have chosen,
unfortunately leads to an extra shift/reduce conflict, which we can
live with.